# Conflicts:
#	.drone.yml
#	installation/sql/mysql/base.sql
#	installation/sql/postgresql/base.sql
#	package.json

+++ administrator/components/com_joomlaupdate/src/Model/UpdateModel.php
@@ -27,7 +27,10 @@ use Joomla\CMS\Plugin\PluginHelper;
 use Joomla\CMS\Updater\Update;
 use Joomla\CMS\Updater\Updater;
 use Joomla\CMS\User\UserHelper;
+use Joomla\CMS\Version;
 use Joomla\Database\ParameterType;
+use Joomla\Registry\Registry;
+use Joomla\Utilities\ArrayHelper;
 
 /**
  * Joomla! update overview Model
@@ -60,7 +63,6 @@ class UpdateModel extends BaseDatabaseModel
 		switch ($params->get('updatesource', 'nochange'))
 		{
 			// "Minor & Patch Release for Current version AND Next Major Release".
-			case 'sts':
 			case 'next':
 				$updateURL = 'https://update.joomla.org/core/sts/list_sts.xml';
 				break;
@@ -71,7 +73,7 @@ class UpdateModel extends BaseDatabaseModel
 				break;
 
 			// "Custom"
-			// TODO: check if the customurl is valid and not just "not empty".
+			// @todo: check if the customurl is valid and not just "not empty".
 			case 'custom':
 				if (trim($params->get('customurl', '')) != '')
 				{
@@ -79,7 +81,9 @@ class UpdateModel extends BaseDatabaseModel
 				}
 				else
 				{
-					return Factory::getApplication()->enqueueMessage(Text::_('COM_JOOMLAUPDATE_CONFIG_UPDATESOURCE_CUSTOM_ERROR'), 'error');
+					Factory::getApplication()->enqueueMessage(Text::_('COM_JOOMLAUPDATE_CONFIG_UPDATESOURCE_CUSTOM_ERROR'), 'error');
+
+					return;
 				}
 				break;
 
@@ -88,6 +92,7 @@ class UpdateModel extends BaseDatabaseModel
 			 * The commented "case" below are for documenting where 'default' and legacy options falls
 			 * case 'default':
 			 * case 'lts':
+			 * case 'sts': (It's shown as "Default" because that option does not exist any more)
 			 * case 'nochange':
 			 */
 			default:
@@ -360,13 +365,39 @@ class UpdateModel extends BaseDatabaseModel
 		$updateInfo = $this->getUpdateInformation();
 		$packageURL = trim($updateInfo['object']->downloadurl->_data);
 		$sources    = $updateInfo['object']->get('downloadSources', array());
-		$headers    = get_headers($packageURL, 1);
+
+		// We have to manually follow the redirects here so we set the option to false.
+		$httpOptions = new Registry;
+		$httpOptions->set('follow_location', false);
+
+		try
+		{
+			$head = HttpFactory::getHttp($httpOptions)->head($packageURL);
+		}
+		catch (\RuntimeException $e)
+		{
+			// Passing false here -> download failed message
+			$response['basename'] = false;
+
+			return $response;
+		}
 
 		// Follow the Location headers until the actual download URL is known
-		while (isset($headers['Location']))
+		while (isset($head->headers['location']))
 		{
-			$packageURL = $headers['Location'];
-			$headers    = get_headers($packageURL, 1);
+			$packageURL = (string) $head->headers['location'][0];
+
+			try
+			{
+				$head = HttpFactory::getHttp($httpOptions)->head($packageURL);
+			}
+			catch (\RuntimeException $e)
+			{
+				// Passing false here -> download failed message
+				$response['basename'] = false;
+
+				return $response;
+			}
 		}
 
 		// Remove protocol, path and query string from URL
@@ -515,8 +546,26 @@ class UpdateModel extends BaseDatabaseModel
 	}
 
 	/**
-	 * Create restoration file and trigger onJoomlaBeforeUpdate event, which find the updated core files
-	 * which have changed during the update, where there are override for.
+	 * Backwards compatibility. Use createUpdateFile() instead.
+	 *
+	 * @param   null  $basename The basename of the file to create
+	 *
+	 * @return  boolean
+	 * @since   2.5.1
+	 * @deprecated 5.0
+	 */
+	public function createRestorationFile($basename = null): bool
+	{
+		return $this->createUpdateFile($basename);
+	}
+
+	/**
+	 * Create the update.php file and trigger onJoomlaBeforeUpdate event.
+	 *
+	 * The onJoomlaBeforeUpdate event stores the core files for which overrides have been defined.
+	 * This will be compared in the onJoomlaAfterUpdate event with the current filesystem state,
+	 * thereby determining how many and which overrides need to be checked and possibly updated
+	 * after Joomla installed an update.
 	 *
 	 * @param   string  $basename  Optional base path to the file.
 	 *
@@ -524,7 +573,7 @@ class UpdateModel extends BaseDatabaseModel
 	 *
 	 * @since  2.5.4
 	 */
-	public function createRestorationFile($basename = null)
+	public function createUpdateFile($basename = null): bool
 	{
 		// Load overrides plugin.
 		PluginHelper::importPlugin('installer');
@@ -556,31 +605,26 @@ class UpdateModel extends BaseDatabaseModel
 		$app->setUserState('com_joomlaupdate.password', $password);
 		$app->setUserState('com_joomlaupdate.filesize', $filesize);
 
-		$data = "<?php\ndefined('_AKEEBA_RESTORATION') or die('Restricted access');\n";
-		$data .= '$restoration_setup = array(' . "\n";
+		$data = "<?php\ndefined('_JOOMLA_UPDATE') or die('Restricted access');\n";
+		$data .= '$extractionSetup = [' . "\n";
 		$data .= <<<ENDDATA
-	'kickstart.security.password' => '$password',
-	'kickstart.tuning.max_exec_time' => '5',
-	'kickstart.tuning.run_time_bias' => '75',
-	'kickstart.tuning.min_exec_time' => '0',
-	'kickstart.procengine' => 'direct',
-	'kickstart.setup.sourcefile' => '$file',
-	'kickstart.setup.destdir' => '$siteroot',
-	'kickstart.setup.restoreperms' => '0',
-	'kickstart.setup.filetype' => 'zip',
-	'kickstart.setup.dryrun' => '0',
-	'kickstart.setup.renamefiles' => array(),
-	'kickstart.setup.postrenamefiles' => false
+	'security.password' => '$password',
+	'setup.sourcefile' => '$file',
+	'setup.destdir' => '$siteroot',
 ENDDATA;
 
-		$data .= ');';
+		$data .= '];';
 
 		// Remove the old file, if it's there...
-		$configpath = JPATH_COMPONENT_ADMINISTRATOR . '/restoration.php';
+		$configpath = JPATH_COMPONENT_ADMINISTRATOR . '/update.php';
 
 		if (File::exists($configpath))
 		{
-			File::delete($configpath);
+			if (!File::delete($configpath))
+			{
+				File::invalidateFileCache($configpath);
+				@unlink($configpath);
+			}
 		}
 
 		// Write new file. First try with File.
@@ -620,9 +664,14 @@ ENDDATA;
 	}
 
 	/**
-	 * Runs the schema update SQL files, the PHP update script and updates the
-	 * manifest cache and #__extensions entry. Essentially, it is identical to
-	 * InstallerFile::install() without the file copy.
+	 * Finalise the upgrade.
+	 *
+	 * This method will do the following:
+	 * * Run the schema update SQL files.
+	 * * Run the Joomla post-update script.
+	 * * Update the manifest cache and #__extensions entry for Joomla itself.
+	 *
+	 * It performs essentially the same function as InstallerFile::install() without the file copy.
 	 *
 	 * @return  boolean True on success.
 	 *
@@ -641,12 +690,15 @@ ENDDATA;
 			return false;
 		}
 
+		// Re-create namespace map. It is needed when updating to a Joomla! version has new extension added
+		(new \JNamespacePsr4Map)->create();
+
 		$installer->manifest = $manifest;
 
 		$installer->setUpgrade(true);
 		$installer->setOverwrite(true);
 
-		$installer->extension = new \Joomla\CMS\Table\Extension(Factory::getDbo());
+		$installer->extension = new \Joomla\CMS\Table\Extension($this->getDbo());
 		$installer->extension->load(ExtensionHelper::getExtensionRecord('joomla', 'file')->extension_id);
 
 		$installer->setAdapter($installer->extension->type);
@@ -713,7 +765,7 @@ ENDDATA;
 		}
 
 		$id = $db->loadResult();
-		$row = new \Joomla\CMS\Table\Extension(Factory::getDbo());
+		$row = new \Joomla\CMS\Table\Extension($this->getDbo());
 
 		if ($id)
 		{
@@ -806,7 +858,7 @@ ENDDATA;
 		ob_end_clean();
 
 		// Clobber any possible pending updates.
-		$update = new \Joomla\CMS\Table\Update(Factory::getDbo());
+		$update = new \Joomla\CMS\Table\Update($this->getDbo());
 		$uid = $update->find(
 			array('element' => 'joomla', 'type' => 'file', 'client_id' => '0', 'folder' => '')
 		);
@@ -841,8 +893,12 @@ ENDDATA;
 	}
 
 	/**
-	 * Removes the extracted package file and trigger onJoomlaAfterUpdate event, which find the updated core files
-	 * which have changed during the update, where there are override for.
+	 * Removes the extracted package file and trigger onJoomlaAfterUpdate event.
+	 *
+	 * The onJoomlaAfterUpdate event compares the stored list of files previously overridden with
+	 * the updated core files, finding out which files have changed during the update, thereby
+	 * determining how many and which override files need to be checked and possibly updated after
+	 * the Joomla update.
 	 *
 	 * @return  void
 	 *
@@ -862,27 +918,30 @@ ENDDATA;
 		$tempdir = $app->get('tmp_path');
 
 		$file = $app->getUserState('com_joomlaupdate.file', null);
-		$target = $tempdir . '/' . $file;
+		File::delete($tempdir . '/' . $file);
 
-		if (!@unlink($target))
+		// Remove the update.php file used in Joomla 4.0.3 and later.
+		if (File::exists(JPATH_COMPONENT_ADMINISTRATOR . '/update.php'))
 		{
-			File::delete($target);
+			File::delete(JPATH_COMPONENT_ADMINISTRATOR . '/update.php');
 		}
 
-		// Remove the restoration.php file.
-		$target = JPATH_COMPONENT_ADMINISTRATOR . '/restoration.php';
-
-		if (!@unlink($target))
+		// Remove the legacy restoration.php file (when updating from Joomla 4.0.2 and earlier).
+		if (File::exists(JPATH_COMPONENT_ADMINISTRATOR . '/restoration.php'))
 		{
-			File::delete($target);
+			File::delete(JPATH_COMPONENT_ADMINISTRATOR . '/restoration.php');
+		}
+
+		// Remove the legacy restore_finalisation.php file used in Joomla 4.0.2 and earlier.
+		if (File::exists(JPATH_COMPONENT_ADMINISTRATOR . '/restore_finalisation.php'))
+		{
+			File::delete(JPATH_COMPONENT_ADMINISTRATOR . '/restore_finalisation.php');
 		}
 
 		// Remove joomla.xml from the site's root.
-		$target = JPATH_ROOT . '/joomla.xml';
-
-		if (!@unlink($target))
+		if (File::exists(JPATH_ROOT . '/joomla.xml'))
 		{
-			File::delete($target);
+			File::delete(JPATH_ROOT . '/joomla.xml');
 		}
 
 		// Unset the update filename from the session.
@@ -1041,19 +1100,16 @@ ENDDATA;
 
 		foreach ($files as $file)
 		{
-			if (File::exists($file))
+			if ($file !== null && File::exists($file))
 			{
-				if (!@unlink($file))
-				{
-					File::delete($file);
-				}
+				File::delete($file);
 			}
 		}
 	}
 
 	/**
 	 * Gets PHP options.
-	 * TODO: Outsource, build common code base for pre install and pre update check
+	 * @todo: Outsource, build common code base for pre install and pre update check
 	 *
 	 * @return array Array of PHP config options
 	 *
@@ -1119,9 +1175,12 @@ ENDDATA;
 		$option->state  = function_exists('json_encode') && function_exists('json_decode');
 		$option->notice = null;
 		$options[] = $option;
+		$updateInformation = $this->getUpdateInformation();
 
-		// Check if configured database is compatible with Joomla 4
-		if (version_compare($this->getUpdateInformation()['latest'], '4', '>='))
+		// Check if configured database is compatible with the next major version of Joomla
+		$nextMajorVersion = Version::MAJOR_VERSION + 1;
+
+		if (version_compare($updateInformation['latest'], (string) $nextMajorVersion, '>='))
 		{
 			$option = new \stdClass;
 			$option->label  = Text::sprintf('INSTL_DATABASE_SUPPORTED', $this->getConfiguredDatabaseType());
@@ -1142,7 +1201,7 @@ ENDDATA;
 
 	/**
 	 * Gets PHP Settings.
-	 * TODO: Outsource, build common code base for pre install and pre update check
+	 * @todo: Outsource, build common code base for pre install and pre update check
 	 *
 	 * @return  array
 	 *
@@ -1233,7 +1292,11 @@ ENDDATA;
 	 */
 	public function isDatabaseTypeSupported()
 	{
-		if (version_compare($this->getUpdateInformation()['latest'], '4', '>='))
+		$updateInformation = $this->getUpdateInformation();
+		$nextMajorVersion  = Version::MAJOR_VERSION + 1;
+
+		// Check if configured database is compatible with Joomla 4
+		if (version_compare($updateInformation['latest'], (string) $nextMajorVersion, '>='))
 		{
 			$unsupportedDatabaseTypes = array('sqlsrv', 'sqlazure');
 			$currentDatabaseType = $this->getConfiguredDatabaseType();
@@ -1267,14 +1330,16 @@ ENDDATA;
 	 */
 	private function getTargetMinimumPHPVersion()
 	{
-		return isset($this->getUpdateInformation()['object']->php_minimum) ?
-			$this->getUpdateInformation()['object']->php_minimum->_data :
+		$updateInformation = $this->getUpdateInformation();
+
+		return isset($updateInformation['object']->php_minimum) ?
+			$updateInformation['object']->php_minimum->_data :
 			JOOMLA_MINIMUM_PHP;
 	}
 
 	/**
 	 * Checks the availability of the parse_ini_file and parse_ini_string functions.
-	 * TODO: Outsource, build common code base for pre install and pre update check
+	 * @todo: Outsource, build common code base for pre install and pre update check
 	 *
 	 * @return  boolean  True if the method exists.
 	 *
@@ -1395,8 +1460,10 @@ ENDDATA;
 		{
 			$decode = json_decode($extension->manifest_cache);
 
-			// Removed description so that CDATA content does not cause javascript error during pre-update check
-			$decode->description = '';
+			// Remove unused fields so they do not cause javascript errors during pre-update check
+			unset($decode->description);
+			unset($decode->copyright);
+			unset($decode->creationDate);
 
 			$this->translateExtensionName($extension);
 			$extension->version
@@ -1455,8 +1522,10 @@ ENDDATA;
 		{
 			$decode = json_decode($plugin->manifest_cache);
 
-			// Removed description so that CDATA content does not cause javascript error during pre-update check
-			$decode->description = '';
+			// Remove unused fields so they do not cause javascript errors during pre-update check
+			unset($decode->description);
+			unset($decode->copyright);
+			unset($decode->creationDate);
 
 			$this->translateExtensionName($plugin);
 			$plugin->version = $decode->version ?? Text::_('COM_JOOMLAUPDATE_PREUPDATE_UNKNOWN_EXTENSION_MANIFESTCACHE_VERSION');
@@ -1494,34 +1563,18 @@ ENDDATA;
 
 				foreach ($updateFileUrls as $updateFileUrl)
 				{
-					$compatibleVersion = $this->checkCompatibility($updateFileUrl, $joomlaTargetVersion);
+					$compatibleVersions = $this->checkCompatibility($updateFileUrl, $joomlaTargetVersion);
 
-					if ($compatibleVersion)
-					{
-						// Return the compatible version
-						return (object) array('state' => 1, 'compatibleVersion' => $compatibleVersion->_data);
-					}
-					else
-					{
-						// Return the compatible version as false so we can say update server is supported but no compatible version found
-						return (object) array('state' => 1, 'compatibleVersion' => false);
-					}
+					// Return the compatible versions
+					return (object) array('state' => 1, 'compatibleVersions' => $compatibleVersions);
 				}
 			}
 			else
 			{
-				$compatibleVersion = $this->checkCompatibility($updateSite['location'], $joomlaTargetVersion);
+				$compatibleVersions = $this->checkCompatibility($updateSite['location'], $joomlaTargetVersion);
 
-				if ($compatibleVersion)
-				{
-					// Return the compatible version
-					return (object) array('state' => 1, 'compatibleVersion' => $compatibleVersion->_data);
-				}
-				else
-				{
-					// Return the compatible version as false so we can say update server is supported but no compatible version found
-					return (object) array('state' => 1, 'compatibleVersion' => false);
-				}
+				// Return the compatible versions
+				return (object) array('state' => 1, 'compatibleVersions' => $compatibleVersions);
 			}
 		}
 
@@ -1641,19 +1694,32 @@ ENDDATA;
 	 * @param   string  $updateFileUrl        The items update XML url.
 	 * @param   string  $joomlaTargetVersion  The Joomla! version to test against
 	 *
-	 * @return  mixed  An array of data items or false.
+	 * @return  array  An array of strings with compatible version numbers
 	 *
 	 * @since   3.10.0
 	 */
 	private function checkCompatibility($updateFileUrl, $joomlaTargetVersion)
 	{
-		$update = new \Joomla\CMS\Updater\Update;
+		$minimumStability = ComponentHelper::getParams('com_installer')->get('minimum_stability', Updater::STABILITY_STABLE);
+
+		$update = new Update;
 		$update->set('jversion.full', $joomlaTargetVersion);
-		$update->loadFromXml($updateFileUrl);
+		$update->loadFromXml($updateFileUrl, $minimumStability);
 
-		$downloadUrl = $update->get('downloadurl');
+		$compatibleVersions = $update->get('compatibleVersions');
 
-		return !empty($downloadUrl) && !empty($downloadUrl->_data) ? $update->get('version') : false;
+		// Check if old version of the updater library
+		if (!isset($compatibleVersions))
+		{
+			$downloadUrl = $update->get('downloadurl');
+			$updateVersion = $update->get('version');
+
+			return empty($downloadUrl) || empty($downloadUrl->_data) || empty($updateVersion) ? array() : array($updateVersion->_data);
+		}
+
+		usort($compatibleVersions, 'version_compare');
+
+		return $compatibleVersions;
 	}
 
 	/**
@@ -1667,7 +1733,7 @@ ENDDATA;
 	 */
 	protected function translateExtensionName(&$item)
 	{
-		// ToDo: Cleanup duplicated code. from com_installer/models/extension.php
+		// @todo: Cleanup duplicated code. from com_installer/models/extension.php
 		$lang = Factory::getLanguage();
 		$path = $item->client_id ? JPATH_ADMINISTRATOR : JPATH_SITE;
 
@@ -1705,6 +1771,65 @@ ENDDATA;
 		|| $lang->load($extension, $source);
 
 		// Translate the extension name if possible
-		$item->name = Text::_($item->name);
+		$item->name = strip_tags(Text::_($item->name));
+	}
+
+	/**
+	 * Checks whether a given template is active
+	 *
+	 * @param   string  $template  The template name to be checked
+	 *
+	 * @return  boolean
+	 *
+	 * @since   3.10.4
+	 */
+	public function isTemplateActive($template)
+	{
+		$db = $this->getDbo();
+		$query = $db->getQuery(true);
+
+		$query->select(
+			$db->qn(
+				array(
+					'id',
+					'home'
+				)
+			)
+		)->from(
+			$db->qn('#__template_styles')
+		)->where(
+			$db->qn('template') . ' = :template'
+		)->bind(':template', $template, ParameterType::STRING);
+
+		$templates = $db->setQuery($query)->loadObjectList();
+
+		$home = array_filter(
+			$templates,
+			function ($value)
+			{
+				return $value->home > 0;
+			}
+		);
+
+		$ids = ArrayHelper::getColumn($templates, 'id');
+
+		$menu = false;
+
+		if (count($ids))
+		{
+			$query = $db->getQuery(true);
+
+			$query->select(
+				'COUNT(*)'
+			)->from(
+				$db->qn('#__menu')
+			)->whereIn(
+				$db->qn('template_style_id'), $ids
+			);
+
+			$menu = $db->setQuery($query)->loadResult() > 0;
+		}
+
+		return $home || $menu;
 	}
 }
